requirements.txt
discord.py == 2.4.0
markovify == 0.9.4
alt-profanity-check == 1.4.1post1
scikit-learn == 1.4.1post1
astra-bot == 1.3.0